% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cohend.R
\name{cohend}
\alias{cohend}
\title{Cohen's D Calculator}
\usage{
cohend(tobj)
}
\arguments{
\item{tobj}{Results of a t.test()}
}
\description{
Calculates Cohen's D effect size for a paired t-test
}
\note{
Based on formula from: http://www.missouristate.edu/assets/rstats/Effect_Size_Calcul  ator_RStats.xlsx
}

