% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataloader.R
\name{dataloader}
\alias{dataloader}
\title{Data Loader}
\usage{
dataloader(folder, expr, ...)
}
\arguments{
\item{folder}{Path to folder where text files exist}

\item{expr}{Regular expression used to grab text files}

\item{...}{Other arguments passed to read_csv}
}
\description{
Loads multiple .csv based on a regular expression and concatenates them into one data.frame
}
\note{
Uses reader and dplyr
}

