% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fordiff.R
\name{fordiff}
\alias{fordiff}
\title{Forward Difference Contrast Maker}
\usage{
fordiff(k)
}
\arguments{
\item{k}{Number of levels in the factor}
}
\description{
Creates a contrast matrix for forward difference coding used in regression
}

