% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modformatter.R
\name{rformatter}
\alias{rformatter}
\title{Format Model Output}
\usage{
rformatter(mod)
}
\arguments{
\item{mod}{A fitted model object.}

\item{stratum}{Which stratum of a repeates measures ANOVA to format.}
}
\value{
Markdown formatted text printed to console.
}
\description{
Takes model output and prints APA formatted results in a RMarkdown document.
}
\details{
Works with output from t.test(), aov(), and nlme().
}

