% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dataloader}
\alias{dataloader}
\title{Data Loader}
\usage{
dataloader(folder, expr, ...)
}
\arguments{
\item{folder}{Path to folder where text files exist}

\item{expr}{Regular expression used to grab text files}

\item{...}{Other arguments passed to read.csv}

\item{cnames}{Character vector of column names to use in produced data.frame}
}
\description{
Loads multiple .csv based on a regular expression and concatenates them into one data.frame
}
\note{
stringsAsFactors is hardcoded to FALSE
}

