% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DataLoader}
\alias{DataLoader}
\title{Data Loader}
\usage{
DataLoader(folder, expr, cnames)
}
\arguments{
\item{folder}{Path to folder where text files exist}

\item{expr}{Regular expression used to grab text files}

\item{cnames}{Character vector of column names to use in produced data.frame}
}
\description{
Loads multiple .csv based on a regular expression files and concatenates them into one data.frame
}

