% Generated by roxygen2 (4.0.2): do not edit by hand
\name{paretasq}
\alias{paretasq}
\title{Partial Eta Squared Caclculator}
\usage{
paretasq(modsum)
}
\arguments{
\item{modsum}{summary object, e.g. summary(anovamod)}
}
\description{
Calculates the partial eta squared effect size statistic for each factor in a repeated measures ANOVA
}
\note{
Formula from: http://jalt.org/test/bro_28.htm
}

